fix: fix reputation score data type

+++ src/main/java/com/campusmov/platform/reputationincentivesservice/reputationincentives/domain/model/aggregates/Valoration.java
@@ -20,7 +20,7 @@ public class Valoration extends AuditableAbstractAggregateRoot<Valoration> {
     private String senderId;
 
     @NotBlank
-    private Integer reputationScore;
+    private Double reputationScore;
 
     @NotBlank
     private String Message;
